* Merge NMUs:
fakechroot (2.6-1.3) unstable; urgency=low

  * Non-maintainer upload to fix the fix for the fix for #422586
  * Only use gcc-multilib on some archs.
  * Fix the readlink function to be ssize_t as it is in newer libc

 -- Sune Vuorela <debian@pusling.com>  Sat, 19 May 2007 17:19:53 +0200

fakechroot (2.6-1.2) unstable; urgency=low

  * Non-maintainer upload.
  * Add build-dependency on gcc-multilib the yada way to avoid FTBFS. 
    (Closes: #422586)

 -- Sune Vuorela <debian@pusling.com>  Sat, 19 May 2007 15:13:28 +0200

fakechroot (2.6-1.1) unstable; urgency=low

  * Non-maintainer upload during BSP.
  * Added build-dependency on gcc-multilib to avoid FTBFS (Closes: #422586).

 -- Mario Iseli <admin@marioiseli.com>  Fri, 18 May 2007 18:17:26 +0200

+++ src/libfakechroot.c
@@ -427,7 +427,7 @@ static int     (*next_open64) (const char *pathname, int flags, ...) = NULL;
 static DIR *   (*next_opendir) (const char *name) = NULL;
 #endif
 static long    (*next_pathconf) (const char *path, int name) = NULL;
-static int     (*next_readlink) (const char *path, char *buf, READLINK_TYPE_ARG3) = NULL;
+static ssize_t     (*next_readlink) (const char *path, char *buf, READLINK_TYPE_ARG3) = NULL;
 static char *  (*next_realpath) (const char *name, char *resolved) = NULL;
 static int     (*next_remove) (const char *pathname) = NULL;
 #ifdef HAVE_REMOVEXATTR
@@ -1967,7 +1967,7 @@ long pathconf (const char *path, int name)
 
 
 /* #include <unistd.h> */
-int readlink (const char *path, char *buf, READLINK_TYPE_ARG3)
+ssize_t readlink (const char *path, char *buf, READLINK_TYPE_ARG3)
 {
     int status;
     char tmp[FAKECHROOT_MAXPATH], *tmpptr;
